evaluation/FactualityPrompts/requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.8
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.11.1
benepar==0.2.0
bitarray==2.6.2
blis==0.7.9
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.8
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
common==0.1.2
confection==0.0.3
cymem==2.0.7
Cython==0.29.32
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fairseq==0.12.2
fever-drqa==1.0.13
filelock==3.9.0
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.1
hydra-core==1.0.7
idna==3.4
importlib-metadata==5.2.0
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
keras==2.11.0
langcodes==3.3.0
Levenshtein==0.20.9
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.8
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.0.6
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
portalocker==2.6.0
preshed==3.0.8
prettytable==3.5.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.13.0
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.2.0
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
thefuzz==0.19.0
thinc==8.1.6
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torch-struct==0.5
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.0
transformers==4.25.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
